+++ tests/conftest.py
@@ -1,3 +1,4 @@
+import ctypes
 import functools
 import inspect
 import platform
@@ -6,8 +7,25 @@ from typing import Any, Callable
 
 import pytest
 
+from pyawaitable.bindings import awaitcallback, awaitcallback_err
+
+try:
+    import _pyawaitable_test
+except ImportError:
+    pytest.exit(
+        "PyAwaitable testing package has not been build! (Hint: pip install tests/extension --no-build-isolation)",  # noqa
+        returncode=-1,
+    )
+
 ITERATIONS: int = 10000
 LEAK_LIMIT: str = "50 KB"
+
+raising_callback = ctypes.cast(
+    _pyawaitable_test.raising_callback, awaitcallback
+)
+raising_err_callback = ctypes.cast(
+    _pyawaitable_test.raising_err_callback, awaitcallback_err
+)
 
 
 def pytest_addoption(parser: Any) -> None:
